refactor:Es6 module
