+++ static/attraction.html
@@ -12,9 +12,7 @@
     <link rel="shortcut icon" href="favicon.ico" type="image/x-icon" />
     <link rel="stylesheet" href="/static/src/css/reset.css" />
     <link rel="stylesheet" href="/static//src/css/all.css" />
-    <script src="/static/src/js/attraction.js" defer></script>
-    <script src="/static/src/js/diolog.js" defer></script>
-    <!-- <script src="/static/src/js/index.js" defer></script> -->
+    <script type="module" src="/static/src/js/attraction.js"></script>
     <title>Attraction</title>
   </head>
   <body>
@@ -83,11 +81,11 @@
         <a class="navigation-brand" href="/">台北一日遊</a>
         <div class="navigation-menu">
           <div class="navigation-collapse">
-            <button class="navigation-btn font-button regular active">
+            <button class="navigation-btn font-button medium active">
               預定行程
             </button>
             <button
-              class="login-btn navigation-btn font-button regular open-dialog"
+              class="login-btn navigation-btn font-button medium open-dialog"
             >
               登入/註冊
             </button>
