+++ static/index.html
@@ -3,12 +3,18 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      rel="preload"
+      href="https://fonts.gstatic.com/s/notosanstc/v35/-nF7OG829Oofr2wohFbTp9iFOisNA_cTyNromxqQuEMQ2wHYwbnmy1R1jDujLebozBXCo2qYhRo.0.woff2"
+      type="font/woff2"
+      crossorigin="anonymous"
+      as="font"
+    />
     <link rel="stylesheet" type="text/css" href="/static/src/css/reset.css" />
     <link rel="stylesheet" type="text/css" href="/static/src/css/all.css" />
     <title>Index</title>
-    <!-- <script src="/static/src/js/api.js" defer></script> -->
-    <!-- <script src="/static/src/js/diolog.js" defer></script> -->
-    <!-- <script src="/static/src/js/list-bar.js" defer></script> -->
     <script type="module" src="/static/src/js/index.js"></script>
   </head>
   <body>
@@ -77,11 +83,11 @@
         <a class="navigation-brand" href="/">台北一日遊</a>
         <div class="navigation-menu">
           <div class="navigation-collapse">
-            <button class="navigation-btn font-button regular active">
+            <button class="navigation-btn font-button medium active">
               預定行程
             </button>
             <button
-              class="login-btn navigation-btn font-button regular open-dialog"
+              class="login-btn navigation-btn font-button medium open-dialog"
             >
               登入/註冊
             </button>
