+++ static/src/css/all.css
@@ -1,25 +1,48 @@
+@import url("https://fonts.googleapis.com/css2?family=Noto+Sans+TC:wght@400;500;700&display=swap");
 * {
   box-sizing: border-box;
 }
 
+@font-face {
+  font-family: "Noto Sans TC";
+  font-style: normal;
+  font-weight: 400;
+  font-display: swap;
+  src: url(https://fonts.gstatic.com/s/notosanstc/v35/-nF7OG829Oofr2wohFbTp9iFOisNA_cTyNromxqQuEMQ2wHYwbnmy1R1jDujLebozBXCo2qYhRo.6.woff2) format("woff2");
+  unicode-range: U+ff78-ff7e, U+ff80-ff86, U+ff89-ff94, U+ff97-ff9e, U+ffb9, U+ffe0-ffe3, U+ffe9, U+ffeb, U+ffed, U+fffc, U+1d7c7, U+1f004, U+1f0cf, U+1f141-1f142, U+1f150, U+1f154, U+1f158, U+1f15b, U+1f15d-1f15e, U+1f162-1f163, U+1f170-1f171, U+1f174, U+1f177-1f178, U+1f17d-1f17f, U+1f192-1f195, U+1f197-1f19a, U+1f1e6-1f1f5, U+1f1f7-1f1ff, U+1f21a, U+1f22f, U+1f232-1f237, U+1f239-1f23a, U+1f250-1f251, U+1f300, U+1f302-1f319;
+}
+* {
+  font-display: swap;
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
+  font-optical-sizing: auto;
+}
+
 h1 {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   font-weight: 700;
   font-size: 1.75rem;
   line-height: 1.5rem;
 }
 
 h2 {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   font-weight: 700;
   font-size: 1.875rem;
   line-height: 2.125rem;
 }
 
 h3 {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   font-size: 1.5rem;
   line-height: 1.5rem;
 }
 
+button {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
+}
+
 .font-btn {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   font-size: 1.188rem;
 }
 .font-btn.bold {
@@ -30,15 +53,18 @@ h3 {
 }
 
 .font-content {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   font-size: 1rem;
 }
 
 .font-body {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   font-size: 1rem;
   line-height: 0.831rem;
 }
 
 .font-category-list {
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   font-weight: 500;
   font-size: 0.875rem;
   font-weight: 500;
@@ -57,7 +83,7 @@ h3 {
 }
 
 body {
-  font-family: "Noto Sans TC", sans-serif;
+  font-family: "Noto Sans TC", "Segoe UI", Frutiger, "Frutiger Linotype", "Dejavu Sans", "Helvetica Neue", Arial, sans-serif;
   overflow-x: clip;
 }
 
@@ -232,7 +258,7 @@ button {
   flex: 1;
   width: 100%;
   padding: 0.813rem 0.625rem;
-  overflow-x: clip;
+  overflow: hidden;
 }
 
 .list-items {
@@ -307,8 +333,8 @@ button {
   position: absolute;
   bottom: 0;
   width: 100%;
-  height: 2.5rem;
-  padding: 0.625rem;
+  line-height: 2.5rem;
+  padding: 0 0.625rem;
   background-color: rgba(0, 0, 0, 0.6);
   color: #ffffff;
   overflow: hidden;
