+++ static/src/js/components/radio.js
@@ -1,11 +1,10 @@
-export function radio() {
-  const bookingForm = document.querySelector(".booking-form");
-  const radio = bookingForm.querySelector(".radio");
-  const morning = document.getElementById("morning");
-  const afternoon = document.getElementById("afternoon");
-  const price = bookingForm.querySelector(".price");
-  const priceAmount = price.querySelector(".price-amount");
+const bookingForm = document.querySelector(".booking-form");
+const morning = document.getElementById("morning");
+const afternoon = document.getElementById("afternoon");
+const price = bookingForm.querySelector(".price");
+const priceAmount = price.querySelector(".price-amount");
 
+function radio() {
   priceAmount.textContent = "新台幣 2000元";
 
   morning.addEventListener("click", () => {
@@ -16,3 +15,5 @@ export function radio() {
     priceAmount.textContent = "新台幣 2500元";
   });
 }
+
+export { radio };
