+++ static/src/js/index.js
@@ -9,6 +9,4 @@ import "./api/tokenApi.js";
 // components
 import "./components/dialog.js";
 import "./components/auth.js";
-import "./components/carousel.js";
-import "./components/radio.js";
 import "./components/listBar.js";
