+++ static/src/js/utils/handleLogout.js
@@ -1,8 +1,10 @@
 import { loginBtn, logoutBtn } from "../variables.js";
-import { updateAuthButton } from "../shared/updateAuthButton.js";
+import { updateAuthButton } from "./updateAuthButton.js";
 
-export function handleLogout() {
+function handleLogout() {
   localStorage.removeItem("authToken");
   updateAuthButton(loginBtn, logoutBtn, false);
   location.reload();
 }
+
+export { handleLogout };
