+++ static/src/js/utils/keywordSearch.js
@@ -2,7 +2,7 @@ import { attractionsApi } from "../api/attractionsApi.js";
 
 const listItem = document.querySelectorAll(".list-item");
 
-export async function keywordSearch() {
+async function keywordSearch() {
   listItem.forEach((item) => {
     item.addEventListener("click", (event) => {
       const searchInput = document.querySelector(".search-input");
@@ -12,3 +12,4 @@ export async function keywordSearch() {
     });
   });
 }
+export { keywordSearch };
