+++ static/src/js/utils/updateAuthButton.js
@@ -1,4 +1,6 @@
-export function updateAuthButton(loginBtn, logoutBtn, isLoggedIn) {
+function updateAuthButton(loginBtn, logoutBtn, isLoggedIn) {
   loginBtn.classList.toggle("active", isLoggedIn);
   logoutBtn.classList.toggle("active", !isLoggedIn);
 }
+
+export { updateAuthButton };
