+++ static/src/scss/components/_attractions-card.scss
@@ -46,8 +46,8 @@
   position: absolute;
   bottom: 0;
   width: 100%;
-  height: 2.5rem; // 40px
-  padding: 0.625rem;
+  line-height: 2.5rem; // 40px
+  padding: 0 0.625rem;
   background-color: $black-opacity-60;
   color: $white;
   overflow: hidden;
