+++ static/src/scss/components/_list-bar.scss
@@ -22,7 +22,7 @@
   flex: 1;
   width: 100%;
   padding: 0.813rem 0.625rem; // 13px 10px
-  overflow-x: clip;
+  overflow: hidden;
 }
 
 .list-items {
